Rename body-builder.js to index.js as entry point to module
